### 2017-03-23 #217 @whedon
[REVIEW]: Ensight4Matlab: read, process, and write data in EnSight Gold format from C++ or MATLAB
**Submitting author:** schmeisser (<a href="http://orcid.org/0000-0002-4400-9635">Andre Schmeißer</a>)
**Repository:** <a href="https://github.com/ITWM-TVFS/Ensight4Matlab" target ="_blank">https://github.com/ITWM-TVFS/Ensight4Matlab</a>
**Version:** v1.0

- [x] **Authorship:** Has the submitting author (schmeisser) made major contributions to the software?
