### 2017-03-23 #217 @Kevin-Mattheus-Moerman
@schmeisser I started the review process. As I said my review will largely focus on the MATLAB integration. I have little experience with C++ so we will find additional reviewers for that component of your work.
I may need some help with installation. I've been following the installation instructions but got stuck (might be doing something wrong) building the EnsightLib library. I use Ubuntu 16.04  64-bit, QMake version 2.01a, Qt version 4.8.7.
If I run:
```
qmake
 qmake-qt4 EnsightLib.pro
make
```
I get the following message:
```
g++ -c -m64 -pipe -std=c++11 -Wall -Wextra -pedantic -O2 -Wall -W -D_REENTRANT -fPIC -DQT_NO_DEBUG -DQT_CORE_LIB -I/usr/share/qt4/mkspecs/linux-g++-64 -I. -I/usr/include/qt4/QtCore -I/usr/include/qt4 -I/usr/local/include -Ilib/.moc -o lib/.obj/ensightlib.o src/ensightlib.cpp
src/ensightlib.cpp:26:19: fatal error: QString: No such file or directory
compilation terminated.
Makefile:261: recipe for target 'lib/.obj/ensightlib.o' failed
make: *** [lib/.obj/ensightlib.o] Error 1
```
