### 2017-03-28 ffe9b3a
Added further MATLAB examples: - Plot of example data set - Extracting a 2D slice of the data set
