### 2017-04-19 #217 @Kevin-Mattheus-Moerman
@schmeisser Thanks. I got a couple of steps further. I'm having some (probably basic) issues with runmex. Btw I also did
`export LD_LIBRARY_PATH=/home/kevin/Desktop/JOSS_review/2017_03_22_Ensight4Matlab/repo/Ensight4Matlab/ensight_lib/lib/:${LD_LIBRARY_PATH}`

Using SETUP.m I set the paths as:

```
PATH settings
  QT(header) 			:	 /usr/include/qt4
  QT(lib) 			:	 /usr/lib/x86_64-linux-gnu/
  Eigen 		:	 /home/kevin/Desktop/JOSS_review/2017_03_22_Ensight4Matlab/3rd/eigen/
  Ensight(header) 	:	 /home/kevin/Desktop/JOSS_review/2017_03_22_Ensight4Matlab/repo/Ensight4Matlab/ensight_lib/include/
  Ensight(lib) 		:	 /home/kevin/Desktop/JOSS_review/2017_03_22_Ensight4Matlab/repo/Ensight4Matlab/ensight_lib/lib/
```

Which seems correct. When I run runmex.m I get a warning (
```
Building with 'g++'.
Warning: You are using gcc version '5.4.1'. The version of gcc is not supported. The version currently supported with MEX is '4.9.x'. For a list of currently supported compilers see:
http://www.mathworks.com/support/compilers/current_release.
> In runmex (line 62)
```
) but it does seem to complete successfully. When I try the examples or
`>> newData = EnsightLib;`
 I get:
```
Warning: The following error was caught while executing 'EnsightLib' class destructor:
Invalid MEX-file '/home/kevin/Desktop/JOSS_review/2017_03_22_Ensight4Matlab/repo/Ensight4Matlab/EnsightMatlab/EnsightLib_interface.mexa64': libEnsightLib.so.1: cannot open shared object file:
No such file or directory.
```
