### 2017-04-20 #217 @schmeisser
@Kevin-Mattheus-Moerman, there seem to be two issues here.

The second error message complains that the file `libEnsightLib.so.1` cannot be found. If you have started MATLAB with the `LD_LIBRARY_PATH` variable set, it should look for the file `/home/kevin/Desktop/JOSS_review/2017_03_22_Ensight4Matlab/repo/Ensight4Matlab/ensight_lib/lib/libEnsightLib.so.1`
You should make sure this file exists. If it doesn't, there has been an error compiling the library. If the file does exist, maybe the paths are not correct.
You can make sure the library has been compiled correctly by building the C++ example first and seeing if it runs, before proceeding to build the MATLAB interface. Let me know if there is any trouble with that.

--------

The first warning message tells us that your system has gcc version 5.4.1 installed, which is not supported by MATLAB.
Unfortunately, things get tricky now!
You need to use a gcc version which is supported by MATLAB, since MATLAB ships with a complete C++ std lib runtime. Depending on the MATLAB version you use, this is unfortunately older than the one on your system.

I have tested the following with MATLAB R2016b, which supports gcc 4.9.

First, install a supported version of gcc:
```bash
> sudo apt-get install g++-4.9
```
Now, you have gcc 4.9 installed but the system will still default to gcc 5.
There are two options to proceed, 1) manually changing the system default or 2) changing the build files to point to gcc-4.9, which is less invasive to your system.

To manually change the system default:
```bash
> sudo rm /usr/bin/gcc /usr/bin/g++
> sudo ln -s /usr/bin/gcc-4.9 /usr/bin/gcc
> sudo ln -s /usr/bin/g++-4.9 /usr/bin/g++
```
Then recompile the C++ library (`make clean; make` in the appropriate folder) and try compiling with MEX again. After you've succeeded, you should restore the original links to your system's default of gcc-5

Alternatively, the following is only slightly more work and doesn't touch the rest of your system:
For the C++ library, edit the Makefile `Ensight4Matlab/ensight_lib/Makefile`, which has been created by `qmake`, and change the first lines reading
```
CC    = gcc
CXX   = g++
```
to
```
CC    = gcc-4.9
CXX   = g++-4.9
```
Then run
```bash
> make clean
> make
```
This should now compile the EnsightLib with a compatible compiler.

Next, on the MATLAB side, edit the file runmex.m. On lines 21-27 it builds a command string for MEX. Here, you need to add the following option to tell it which gcc to use:
```
' -v GCC= gcc-4.9 '
```
Make sure to include spaces before and after this, so that the new option is separated from the existing options.
In the output of runmex you should now see gcc-4.9 being used and the warning about an incompatible compiler should disappear.

--------

If everything works, the 4 scripts in the examples folder should successfully run now.
