### 2017-07-11 #217 @Kevin-Mattheus-Moerman
@schmeisser If I alter `runmex.m` with:

`        mex_str = ['mex ',OPTS,...
                   'CXXFLAGS="\$CXXFLAGS -std=c++11"', ...
                   ' -I',QT_INCLUDE_PATH,' -I',QT_INCLUDE_PATH,'/QtCore',...
                   ' -I',EIGEN_PATH,...
                   ' -I',ENSIGHT_INCLUDE_PATH,...
                   ' -L',QT_LIB_PATH,' -lQtCore -L',ENSIGHT_LIB_PATH,' -lEnsightLib ',...
                   ' -v GCC= gcc-4.9'];`

> Error using mex
> /home/kevin/Desktop/JOSS_review/Ensight4Matlab/EnsightMatlab/gcc-4.9 not found; check that you are in the correct current folder, and check the spelling of '/home/kevin/Desktop/JOSS_review/Ensight4Matlab/EnsightMatlab/gcc-4.9'.
>
> Error in runmex (line 63)
>             eval([mex_str,' mexfiles/EnsightLib_interface.cpp']);

That made me think I should specify the path to gcc-4.9 so I tried adding  `'-v GCC= /usr/bin/gcc-4.9 '`, but then I get the erro:

>
> Error using mex
> Unknown file extension '.9'.
>
> Error in runmex (line 63)
>             eval([mex_str,' mexfiles/EnsightLib_interface.cpp']);

Am I doing something wrong? Should it be `' -v ''GCC= /usr/bin/gcc-4.9'' '`
