### 2017-07-11 #217 @Kevin-Mattheus-Moerman
With that change `runmex.m` does run it seems, but when I try: `newData = EnsightLib;` I get:

> Warning: The following error was caught while executing 'EnsightLib' class destructor:
> Invalid MEX-file '/home/kevin/Desktop/JOSS_review/Ensight4Matlab/EnsightMatlab/EnsightLib_interface.mexa64':
> /usr/local/MATLAB/R2016b/bin/glnxa64/../../sys/os/glnxa64/libstdc++.so.6: version `GLIBCXX_3.4.21' not found (required by
> /lib/libEnsightLib.so.1).
>
> Error in EnsightLib/delete (line 221)
>       EnsightLib_interface('interface', 'delete', [1, 0, 5], this.getObjectHandle());
>
> Error in EnsightLib (line 172)
>     function this = EnsightLib(filename)
> > In EnsightLib (line 172)
> Invalid MEX-file '/home/kevin/Desktop/JOSS_review/Ensight4Matlab/EnsightMatlab/EnsightLib_interface.mexa64':
> /usr/local/MATLAB/R2016b/bin/glnxa64/../../sys/os/glnxa64/libstdc++.so.6: version `GLIBCXX_3.4.21' not found (required by
> /lib/libEnsightLib.so.1).
>
> Error in EnsightLib (line 180)
>             this.EnsightObjectHandle = EnsightLib_interface('interface', 'create', valueArray);
