### 2017-07-11 #217 @schmeisser
Sorry, I think there was a typo in my last message.
It should be ' -v GCC=gcc-4.9 ', i.e. there should be *no space* after (or before) the equals sign.

The second error you see about "version `GLIBCXX_3.4.21' not found" is a symptom of the original problem, compiling with a gcc version and thus a standard library that is newer than the one supported by MATLAB. This should not occur if you fix my typo and try again compiling with gcc 4.9.
