### 2017-07-12 59b99b8
Fixed path truncation bug
