### 2017-07-12 #217 @schmeisser
The LD_PRELOAD trick sounds like it still used the newer gcc to compile, but you forced Matlab to load a newer library version. As far as I know, that isn't officially supported, but seems to work.

The error message seems to be caused by a truncation bug. As a workaround you can try to `cd` into the `data` directory and use a relative path (just `jet.encas`) instead of an absolute one. In the meantime, I will upload a patch to fix the bug.
