### 2017-07-12 #217 @Kevin-Mattheus-Moerman
Yeah if I cd into the data directory I can run this (note I had to add jet/ as I unzipped the data into the folder jet)
`jet = EnsightLib('jet/jet.encas')`
I can confirm this works as well as the examples and `plot_jet.m`.
