### 2017-07-13 #217 @Kevin-Mattheus-Moerman
I'll now spend some more time reviewing the code, examples, documentation, ,and functionality
