### 2017-07-13 #217 @Kevin-Mattheus-Moerman
@cMadan I'm at the stage where I can reproduce the basic functionality on the MATLAB side of things. I'll formulate some comments shortly on the MATLAB integration and will proceed to tick the boxes above. I think it would be good to also have a reviewer for the C++ elements of this submission. @cMadan do you think it is possible to recruit an additional reviewer?
