### 2017-07-13 #217 @daviddoria
@Kevin-Mattheus-Moerman I'd be happy to look at it, but I'll be travelling through the end of the month. If it can wait until mid-August you can count me in, but if you need it sooner I'll have to pass.
