### 2017-07-17 #217 @cMadan
@daviddoria - that sounds good to me, thank you!

@Kevin-Mattheus-Moerman - are you happy to sign off on the MATLAB portion then, or is there anything else that should be cleared up in the meantime? I know you had some issues with compiling the mex files, so perhaps the installation instructions need to be improved? How about the automated tests and API documentation? Thanks again for all of your work in reviewing this!
