### 2017-07-19 #217 @schmeisser
If there are suggestions on improving the installation instructions regarding the Matlab side of things, I'll be happy to add more information.
