### 2017-07-19 #217 @Kevin-Mattheus-Moerman
@schmeisser I'm in the process of formulating my suggestions. I'm a bit swamped this week but should be able to get to done by Friday
