### 2017-08-14 #217 @cMadan
@Kevin-Mattheus-Moerman - it would be great if you could provide your suggestions soon :)

@daviddoria - are you still available to review the C++ portion of this submission?
