### 2017-08-14 #217 @daviddoria
@cMadan Very sorry for the delay. I will get this done Wednesday evening. First thoughts: WONDERFUL USAGE DOCUMENTATION!! I haven't tried it yet, but I'm very hopeful that it will Just Work™ :)
