### 2017-08-14 #217 @schmeisser
@daviddoria Since the Matlab portion didn't Just Work™, mostly because of building/compiler issues, I'll gladly try to help out if you should run into any problems, and adapt/add to the documentation as necessary.
