### 2017-08-16 #217 @daviddoria
@cMadan @Kevin-Mattheus-Moerman  @schmeisser

Hi guys,

Overall, I am very happy with this - nice work! I did a "code review" of the whole repository by forking, creating a branch based off of the first commit in your repository, and submitting a pull request (a cool trick I learned here: http://astrofrog.github.io/blog/2013/04/10/how-to-conduct-a-full-code-review-on-github/)

You can find my comments here:
https://github.com/daviddoria/Ensight4Matlab/pull/1#pullrequestreview-56735748

Most things are pretty minor, but I'll repeat the two "big ticket" items here:

-----
1) There does not appear to be a single test - this is very frightening. Even making a trivial sample data set that you read and confirm the contents of would make me feel a little better. But particularly since you're using custom things like the bounding box class, etc. I would never trust it without a bunch of tests.

2) I'd love to see a non-GUI example, basically to demonstrate how easy it is to use the major pieces of the library (the reader and writer). A "5 line" main like:

object = read(file)
do something
write(object, newfile)

would be a welcome addition (and serve as a basic "test" as well (at least an "execution test")).
-----

Please let me know if you have any questions, need anything clarified, etc.

Thanks!

David
