### 2017-08-22 #217 @cMadan
Thanks for the detailed review @daviddoria! I really appreciate your thoroughness here (and for the link re: code review).

@schmeisser, let me know after you've had a chance to revise the project based on this feedback.
