### 2017-08-28 #217 @schmeisser
Thanks for the thorough review, I'll get back to you once I've found the time to work through the individual items.
