### 2017-09-26 205dcd1
JOSS review: - Moved write functions from classes Ensight*Writer to namespace Ensight::Writer - Added more documentation - Some formatting changes - Minor changes
