### 2017-09-27 2da3d11
Added more info about installation / setting LD_LIBRARY_PATH
