### 2017-09-27 #217 @schmeisser
Sorry for the long delay, I've been a bit busy.

I've now added the suggested wording about LD_LIBRARY_PATH to the README.md and incorporated the suggestions from your code review, i.e. changed the writer to be in namespace Ensight::Writer, added more documentation, and some other small changes.

The suggested tests are on the way and will be added soon, and I'll also add some minimal C++ examples to go alongside the GUI example, shortly.
