### 2017-10-06 6f33d9b
Improved barycentric coordinate computation
