### 2017-10-17 1248611
Examples for reading data set, interpolating variable values
