### 2017-10-17 249240a
Example for writing a data set, reference data files
