### 2017-10-18 cce3bf1
Add tests
