### 2017-10-18 #217 @schmeisser
I have added some tests in the directory `ensight_lib/test` and 3 short command line examples in `ensight_lib/examples`.
For testing that writing works correctly the `write_file` example can be used, as it does exactly that. It includes reference data files of what the result should be. The output produced by the example can be verified to be correct by using a standard file `diff` tool to compare it to these reference files.
