### 2017-10-19 #217 @daviddoria
@schmeisser Thanks - we're really close! A few things:

1) I'm not familiar with `qmake`, but the only way I could get the tests and examples to build was to copy the same Eigen directory into `INCLUDE_PATH` into both `EnsightLibTest.pro` and each of the examples .pro files (write_file.pro, etc.). Coming from CMake, that seems super annoying. Is there not a way to tell QMake at the top level "build everything below me" with the same settings? (the equivalent to `add_subdirectory()` in cmake)?

2) Again not being familiar with QMake I had to poke around a little bit to figure out what happened/what I was supposed to do after I built EnsightLibTest.pro. Obviously, I found the build/ directory and realized I just needed to run the executable EnsightLibTest, but let's add that to the README instructions (especially when the header is "Running tests").

3) The tests did not all pass for me:

```
********* Finished testing of EnsightConstantTests *********
********* Start testing of EnsightReaderTests *********
Config: Using QTest library 4.8.7, Qt 4.8.7

FAIL!  : EnsightReaderTests::ReadCaseFile_SuccessfulCaseFileRead_CaseFileCorrectlyParsed() 'successfulRead' returned FALSE. ()
   Loc: [ensightreadertests.cpp(79)]

FAIL!  : EnsightReaderTests::QStringRead_SuccessfulCaseFileRead_CaseFileCorrectylParsed() '(bool) createdEnsightObj' returned FALSE. ()
   Loc: [ensightreadertests.cpp(319)]

Totals: 26 passed, 2 failed, 0 skipped
```
