### 2017-10-27 ebd54b2
Added warning if reference data files are not found.
