### 2017-11-13 94c377a
Configuration for includes moved from .pro files to EnsightLibConfig.pri
