### 2017-11-13 72d6e30
Updated README about EnsightLibAndTests.pro
