### 2017-11-13 #217 @schmeisser
1. You're right, I've searched a bit and found a much better way to do this. Two changes: First, there's now a top-level file `EnsightLibAndTests.pro` that simply builds all other .pro files (library, tests, and examples. Didn't want to make the file name too long, though). If required, the individual pro files can of course still be compiled separately. Second, the `INCLUDE_PATH` is now defined in file `ensight_lib/EnsightLibConfig.pri`, which is referenced from all other .pro files, so it needs to be changed only once. This does mean that the path must be absolute now, though, as the referencing .pro files are in different directories.

2. That was quite non-obvious, and also not consistent with the examples. I've simply changed it to put the binary in the same directory, the same way the examples do it. I've also adapted the README to point it out. Hopefully it's clearer now.

3. This is probably because the reference files cannot be found.
Maybe there's a problems with the symlinks to the file there, although I tested it and they work for me. Otherwise, maybe you just did not unpack the example data in `data/jet.tar.bz2`. I've added an error message in the tests to point out if files simply cannot be found. I've also changed the README to make this point stick out more.
