### 2017-11-15 #217 @daviddoria
@schmeisser  Great, the process went smoothly this time!

A few final things (from me):

1) for paper.md - is this the desired format of the Markdown? https://github.com/ITWM-TVFS/Ensight4Matlab/blob/master/paper/paper.md

I see a "code block" at the top with the author list/affiliation and then the "rendered" Summary and References sections below it.

2) "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I do not see anything about this in the paper.md

3) "Version: Does the release version given match the GitHub release (v1.0)?"
I don't see any mention of Releases or Versions at all.

David
