### 2017-11-15 #217 @Kevin-Mattheus-Moerman
In relation to 1), yes that is what the paper looks like at this stage.
