### 2017-11-16 #217 @schmeisser
I'm not sure about these.
I haven't set up any community guidelines, I would be fine if potential contributors contact me directly over github. Is that even possible?
The initial release version would be 1.0, I guess. Where do I have to mention this?
