### 2017-11-16 #217 @Kevin-Mattheus-Moerman
Hi @schmeisser. I can offer some information on community guidelines.

- Github has a useful [site](https://opensource.guide/building-community/) on setting up community and contributing guidelines. They also link [here](https://help.github.com/articles/setting-guidelines-for-repository-contributors/) for info on contributer guidelines
- You can also check [my personal project](https://github.com/gibbonCode/GIBBON). The readme there links to a contributing and a conflict of interest document. I recommend adopting something similar.
- Here is a [link](https://www.contributor-covenant.org/version/1/4/code-of-conduct.html) you can use to base your code of conduct document on

Let me know if you have any other questions.
