### 2017-11-16 #217 @Kevin-Mattheus-Moerman
About versions. I see you have not minted a release yet in your repository. As part of the JOSS acceptance process you will create a release (the version accepted in JOSS) and archive it for instance using ZENODO  (See [here](https://guides.github.com/activities/citable-code/) to set this up but you can wait until formal acceptance). When you create your first release make sure the version is labeled as `v1.0` (completely up to you so make sure you set it correctly).
