### 2017-11-16 #217 @Kevin-Mattheus-Moerman
@schmeisser in relation to your paper. Please make sure your citations are in [rMarkdown format](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html).
For instance, instead of writing [3] this citation should be written as: [@gramsch2014simulation], where the keyword after the @ symbol refers to the paper.bib file entries.

Let me know if I can help.
