### 2017-11-17 #217 @schmeisser
@Kevin-Mattheus-Moerman Thanks, I'll check out your links on writing a contribution guideline. I guess for the beginning it's fine to keep it simple.
Can you please check if the updated paper is formatted as required?
