### 2017-11-17 #217 @Kevin-Mattheus-Moerman
@schmeisser Great, yeah you can keep it simple and simple copy past stuff from those suggested ones.
The paper seems appropriately formatted now. If there are any other issues we'll find them at the stage when the PDF is generated.
