+++ CONTRIBUTING.md
@@ -1,0 +1,22 @@
+Thank you for taking your time to contribute!
+
+# Code of Conduct
+To keep it simple: please be considerate and respectful, behave mannerly and use your common sense.
+
+# Contribution Guidelines
+## Reporting issues
+
+- **Search for existing issues.** Please check to see if someone else has reported the same issue.
+- **Share as much information as possible.** Include operating system and version, as well as MATLAB version. Also, include steps to reproduce the bug.
+
+## Pull requests
+- Keep pull requests simple and small, do not pollute them with unintended changes.
+- If your PR resolves an issue, include **closes #ISSUE_NUMBER** in your commit message (or a [synonym](https://help.github.com/articles/closing-issues-using-keywords/)).
+
+## Code Style
+
+Not all current code follows the conventions below but these will be followed for future developments. 
+- `lowerCamelCase` variables and functions/methods.
+- `UpperCamelCase` classes.
+- Use descriptive variables names when possible, e.g. `partName` instead of `pn`. Short variable names are okay for local with small scope, e.g. loop variables. 
+- Indent code using [Allman style](https://en.wikipedia.org/wiki/Indentation_style#Allman_style). If surrounding code significantly differs from this, consider using the style predominantly used in the file instead.
