### 2017-12-04 #217 @schmeisser
@Kevin-Mattheus-Moerman I have copied some parts of your contribution guideline as you suggested, and also added a section about code of conduct and how to seek support.
