### 2017-12-04 #217 @arfon
> @whedon generate pdf

A couple of obvious(?) errors with the `paper.md` - the title will need quoting in double quotes as there's an apostrophe in the title.

One of the authors is missing an affiliation.
