### 2017-12-06 5d0cc97
Fix quoting in paper.md
