### 2017-12-06 #217 @arfon
> The date in `paper.md` still says 15 March 2017. Do I need to update this to the current date?

That date doesn't actually show up in the paper anywhere so I think you can leave it as-is.
