### 2017-12-06 #217 @Kevin-Mattheus-Moerman
@arfon I was going to suggest to this author to fix the link to their repo in the paper to be a hyperlink (see last line in paper). But actually I wonder if we should remove this text altogether. I know we ask people to mention the repository link in the paper. But since our PDF article now has a link rendered in it in a clear spot I think there is no need to refer to it in the text anymore. Thoughts?
