### 2017-12-06 #217 @arfon
> But since our PDF article now has a link rendered in it in a clear spot I think there is no need to refer to it in the text anymore. Thoughts?

Yes, I've generally been asking authors to remove links like this as we automatically create a prominent link in the JOSS papers.
