### 2017-12-06 da5ef40
Removed repository link from paper.md
