### 2017-12-07 #217 @arfon
@cMadan & @Kevin-Mattheus-Moerman - is this good to accept now?
