### 2017-12-07 ead9213
Remove pronouns from title in paper.md
