### 2017-12-07 6570057
Acknowledge OpenMesh in viewer example
