### 2017-12-07 #217 @schmeisser
@arfon I have made a minor edit to the title to remove and minted a v1.0.0 release.
