+++ README.md
@@ -6,6 +6,7 @@ Ensight4Matlab let's you read, process, and write files in EnSight&reg; Gold for
 Data in the Ensight Gold format represents a 3D mesh and variable fields defined over the domain of this mesh. The mesh and/or variables may be either static in time or time-varying (transient). The Ensight Gold format is used to save e.g. CFD and CAE data in research and industrial applications, and widely supported by many software tools. It is specified by CEI software, Inc., see http://www.ceisoftware.com for more details.<br>
 This packages provides a C++ based library to process e.g. CFD data sets in EnSight Gold format, which you can directly link (statically or dynamically) to your application. Additionally, it provides a language binding for MATLAB to easily use the C++ library in your MATLAB scripts, e.g. for analysis and rapid prototyping.
 
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1095608.svg)](https://doi.org/10.5281/zenodo.1095608)
 
 Table of Contents
 =================
