### 2017-12-08 #217 @cMadan
@arfon, I think we're all set to accept!

@Kevin-Mattheus-Moerman and @daviddoria, thank you both for your thorough reviews of this submission!
